Implement input field and state for controlling area search minutes.
